table view click makes key window, so we get same behavior as IKImageBrowserView, and we can use quickLook on selected item.  Issue 213.
The client app, after showing the window, makes the panel key.  Our stand-alone app will probably want to do something similar.  This is nice to not have this code in the framework at all!

Set development language to 'en' per http://eschatologist.net/blog/?p=224  ... to be able to migrate to 2-letter language codes, and hopefully so each checkin doesn't change things in the project file!