bugfix from Jörg:
I got a report that there is a problem with the reload functionality on iPhoto nodes in the current version. One will always get 0 objects displayed on automatic or manual reload.

I tracked it down to a missing setup of the attributes ivar in the new node from the old node in method nodeWithOldNode:options:error: of IMBiPhotoParser. This is essential since we now utilize this ivar.

--HG--
extra : convert_revision : svn%3A0a3f77c4-9235-0410-95f4-d138c4647c51/trunk%401401