Checked in Jörg's beta 3 of skimming. It's pretty close.  We may need to adjust the linking a bit.  Also, I think we should probably discuss the best course for automatically changing view type; I think it is probably better to not adjust things automatically.
This is "close enough" that I'm checking this in; it will make it easier to deal with smaller change increments at this point.

Release Notes:


Skimming Beta 3 Release Notes (2011-02-17):
-------------------------------------------

1. Hooks into the framework
Added a hook into the framework so that its host app can set the image browser background layer. Added some sample code to the test application to show how to provide a customized image browser cell and image browser background layer.

2. Fixed accidental redefinition of "delegate" from base class in IMBSkimmableObjectViewController

3. Accompanied patch file is now unedited, i.e. Xcode project file diffs are included.

4. Known Issues:
  - event view loses zoom factor when left and then returned to


Skimming Beta 2 Release Notes (2011-02-16):
-------------------------------------------

1. Fixed Safari bookmark parser infinite loop on update events from Safari

2. Fixed problem with image browser cell attached CALayers which were sometimes drawn off position

3. Event related images are now drawn "aspect fill" (no white/black space anymore)

4. Regarding 2. and 3. I revised the implementation for drawing event cells from using CALayer to using Core Graphics. This afterall makes it possible to draw "aspect fill" into image browser cells. Furthermore, drawing only portions of an image is an indispensible prerequisite for implementing iPhoto Faces.

5. Skimming Performance
Fixed performance problems when skimming on events. This was attributed to delay factors implemented in IMBObjectController methods dealing with image browser reloads. I adjusted those delays for the skimming controller to more appropriate values.

6. Licensing and Authoring Info
Added proper headers for newly introduced files.


Skimming Beta 1 Release Notes (2011-02-14):
-------------------------------------------

1. Modification of IMBNodeObject
IMBNodeObject now has a new property "representedNodeIdentifier". With this the overloading of IMBObject's property location (which was used as a path (IMBObject) or an IMBNode* (IMBNodeObject)) is now eliminated. All code throughout the framework that accessed .location to set or retrieve an IMBNode is adapted accordingly. For lack of proper testing environments I did not test all libraries. Specifically, I did not test Firefox view, Lightroom 1 and 3 views. Also, it seems that the Safari browser now is caught in an infinite loop of trying to repopulate himself whenever it gets triggered by an update event from Safari (any hints welcome).

2. Skimming and OS X 10.5
Using the OS X ImageKit (specifically the IKImageBrowserView) I believe it is not possible to provide proper skimming without utilizing Apple private API. Skimming may work rudimentarily in 10.5 but we may want to decide to exclude 10.5 from skimming altogether. I yet have not tested the code on 10.5 at all (for lack of environment).

3. Skimming on iPhoto Movies
This seems to be excessively slow and I have not really looked into it yet.

4. Licensing and Authoring Info
Proper information headers for newly introduced files have yet to be included.

5. NSLog()
There are still some logging statements in the code which will be removed for release.