From Jörg:
Skimming Release Candidate Release Notes (2011-02-24):
------------------------------------------------------

Note: This release also includes all Beta 4 changes (and hence release notes) since Beta 4 changes were not committed to the repository (due to a change conflict with change 1515).

1. Fixed "array out of bounds" exception when trying to skim on a filtered subset of events

2. Fixed loss of current zoom factor in event view when leaving the view and then returning back to

3. Fixed crash on combo view selection when viewing movie events (also see item 9.)

4. Fixed memory leak when skimming on events

5. Fixed crash on 10.5.8 caused by revision 1515 (use of MDItemCreateWithURL() on 10.5)

6. Fixed display of non qualified objects while skimming (erroneously included images on movie skimming and vice versa). With this I think we may close issue 244.

7. Added some metadata for event display in combo view and tooltip (number of events, event date). With this I think we may close issue 245.

8. Delegate Protocols
Moved imageBrowserBackgroundLayerForController: method from IMBLibraryController delegate protocol to IMBObjectViewController delegate protocol where it rather belongs to (thanks to Peter for pointing this out)

9. Support for 10.5
I tested the sample application on 10.5.8 (x86_64). It runs quite well. To be mentioned:
- Skimming is not as snappy as on 10.6
- No layer support for image browser view (handled in framework code)
- Sporadic console error messages from Image Kit about a changed UID that don't seem to affect functionality

!!! The following is important !!!
10. Refactored Image Representation Support through Quick Look for IMBObjects:

Since the combo view in the movies library uses Quick Look to display image representations of movies I had to support this too for movie events. Unfortunately, the Quick Look image creation support was implemented solely in IMBMovieObject although it did not leverage any IMBMovieObject specific behavior.

To support image creation through Quick Look in IMBiPhotoEventNodeObject I either had to copy the functionality to this class or move it up the inheritance hierarchy to the first common ancestor of IMBMovieObject and IMBiPhotoEventObject which is IMBObject. After discussing this with Peter we agreed that it is best to move the functionality to IMBObject (which is what I did).

With this we gain the opportunity to more easily integrate any file formats into iMedia that are supported by Quick Look image creation.

Also with this IMBMovieObject degrades to an empty stub. I consequently also translated all references to IMBMovieObject to IMBObject. The few existing code fragments "if object is kind of IMBMovieObject..." were easily translated to decisions based on the image representation type. Thus, there are no more references to IMBMovieObject inside the framework anymore - I just didn't dare to remove the class itself.

Of course there is a slight chance that host applications still count on the instantiation of IMBMovieObjects which are no longer instantiated. Of course, we could still instantiate IMBMovieObject instances but would then lose the opportunity to remove the class altogether.

Makes sense?

11. Removed ephemeral NSLog statements

Known Issues: none


Enjoy!