Prevent the propagation of errors encountered in the "inner loop" of populateNode:options:error. These errors were being vended to the client as zombies because they are allocated and autoreleased by AppKit in the context of our local autorelease pool.
I suspect this was not noticed more often because of the luck of another buggy behavior here: the
inner loop loading of all these subnode contents was consistently overwriting the old value of
error, so that in most cases if an error was encountered during the inner loop, a successful
load of a sibling folder after the error folder would cause the error to be overwritten with nil.

I am fixing this by avoiding propagation of these "inner loop" errors altogether. This makes
sense because even if they hadn't been vended as zombies, vending them at all sets the error status
for the whole populateNode call, causing the folder outline to completely omit the contents of the
folder.

With these fixes in place, there will never be a zombie error vended, because the only error propagated
is one that is allocated before any autorelease pool is established. Furthermore, the error
will only be propagated when the specific outer node being loaded has an error reading contents. The
net effect of this to the user is that when you click on a folder that CONTAINS unreadable nodes, you
still see all the files. But when you click one of the subfolders that is itself unreadable, only
then do you get a user-propagated error and expected empty contents display.