Do not query master resource file for extra metadata in Lightroom if we are sandboxed
This is because this file may very well reside in a not yet entitled location (e.g. on an additional storage device) thus prompting the user for granting even more entitlements involving the need to chose an appropriate to be entitled directory carefully.

NOTE: until we get a chance to provide the creation date from Lightroom's metadata database it will be missing in the browser.