Add completion handling for parser specific tasks all the way up to actions initiated by IMBNodeViewController; also improve error handling;
Once Facebook access token expires we need to automatically re-authenticate and retry an issued request to Facebook. Since authentication must be done in app process and on main thread (sandboxed cookies and webview involved) a completion handler at the level of IMBNodeViewController must be established. This completion handler will handle automatic re-authentication and re-issuance of the action that failed (like populateNode:, reloadNode:, or loadThumbnails).