Ensure that IMBFacebookParser instance always references the current PhFacebook object (XPC services may get terminated by the system)
On 10.7 XPC services may get terminated after a couple of seconds of inactivity. This resulted in a newly instantiated IMBFacebookParser upon next request to the service. This instance had no reference to the current PhFacebook object since it was only set on login.

We now use the IMBFacebookParserMessenger to hold on to the current PhFacebook object and re-set it on the IMBFacebookParser whenever a new request is issued to the parser.